Add TTL to in-progress markers
Add a finite lifetime to markers so that they can expire and prevent
permanently blocking the creation or recreation of a digest. This
patch adds a timestamp as the contents of the marker file and adjusts
the behaviour of the InProgress wrapper to check the timestamp whenever
it finds a marker.

Even if the marker is never deleted this patch will enable the system
to process a new request for a previously in-progress digest in the
event that something went wrong with the digest creation.

Note, we are not using the S3 TTL feature because the granularity is
only in days. We would have to wait a minimum of 24 hours before we
could attempt another digest.